requirements.txt
aiohttp==1.0.5
async-timeout==1.2.1
chardet==3.0.4
discord.py==0.16.8
multidict==2.1.6
psycopg2==2.7.1
websockets==3.3
requests==2.18.4
